Disable warning when nix on PATH
Currently the stack configuration explicitly disables nix integration. However
Stack is configured in such a way, that when nix integration is disabled but nix
executable is available on the PATH, the stack will issue a warning each time it
is being called.

The `notify-if-nix-on-path: false` disables this.

Fun fact, stack recognizes if user is using NixOS Linux distribution and
disables this warning for NixOS users

Yes.

Users without nix will not see the difference.
Users with nix will no longer see the warning.
Users on NixOS will not see any difference.