Add messageTypeWeights and use it value calculating a score for given message type
The reliability and latency importance may vary depending on the message
type. For instance for small 'heartbeat-like' messages it might be important for
them to be delivered quickly, and it's smaller issue if they do not always
arrive. On the other hand, for the conseuns messages it is crucial that they are
being delivered even if it takes longer time