Add check of the invariant that calculated peer score is in between [0.0, 1.0]
Please note that we did not bring the old tests back here to live as this was a
challange of its own. Instead we introduced new-test folder.

Evenatually test and new-test should be merged - by either bring P2P tests back
to life or removing them completely.

For now this workaround with a new-test folder should suffice